refactor: Move common composables to common file
