+++ app/src/main/java/com/github/onlynotesswent/ui/search/Search.kt
@@ -25,11 +25,11 @@ import androidx.compose.ui.text.input.TextFieldValue
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import com.github.onlynotesswent.model.note.NoteViewModel
+import com.github.onlynotesswent.ui.NoteItem
 import com.github.onlynotesswent.ui.navigation.BottomNavigationMenu
 import com.github.onlynotesswent.ui.navigation.LIST_TOP_LEVEL_DESTINATION
 import com.github.onlynotesswent.ui.navigation.NavigationActions
 import com.github.onlynotesswent.ui.navigation.Screen
-import com.github.onlynotesswent.ui.overview.NoteItem
 
 /**
  * Displays the search screen where users can search notes by title.
